+++ tests/test_workflow.py
@@ -22,7 +22,18 @@ except ImportError:
 
 
 CFG_FILE_BACKWARD = Path("tests/test_data/config_rhine.yaml")
+CFG_FILE_BACKWARD_PL = Path("tests/test_data/config_rhine_pl.yaml")
 CFG_FILE_FORWARD = Path("tests/test_data/config_west_africa.yaml")
+
+REGRESSION_DATA_MODEL_LAYERS = (
+    "tests/test_data/verify_output/preprocessed_data/2022-08-31_fluxes_storages.nc"
+)
+REGRESSION_DATA_PRESSURE_LAYERS = (
+    "tests/test_data/verify_output/preprocessed_data/2020-01-01_fluxes_storages.nc"
+)
+
+RTOL = 1e-5  # 0.001 %
+ATOL = 1e-7  # kg/m2 == 1e-7 mm (accumulated per timestep)
 
 
 ## First we define fixtures to help us write the tests more concisely
@@ -52,6 +63,11 @@ def preprocessed_dir(temp_directory: Path) -> Path:
 
 
 @pytest.fixture(scope="session")
+def preprocessed_dir_pl(temp_directory: Path) -> Path:
+    return temp_directory / "preprocessed_pl"
+
+
+@pytest.fixture(scope="session")
 def figures_directory(temp_directory: Path) -> Path:
     fig_dir = temp_directory / "figures"
     fig_dir.mkdir()
@@ -71,6 +87,23 @@ def test_config_backward(
     cfg.preprocessed_data_folder = preprocessed_dir
     cfg.output_folder = output_dir_backtrack
     tmp_config = temp_directory / "config_rhine.yaml"
+    cfg.to_file(tmp_config)
+    return str(tmp_config)
+
+
+@pytest.fixture(scope="session")
+def test_config_backward_pl(
+    temp_directory: Path, preprocessed_dir_pl: Path, output_dir_backtrack: Path
+) -> str:
+    """Return the path to the config for the preprocess, backward, and visualize tests.
+
+    The processed and output data folders of this config are modified to use temporary
+    directories.
+    """
+    cfg = Config.from_yaml(CFG_FILE_BACKWARD_PL)
+    cfg.preprocessed_data_folder = preprocessed_dir_pl
+    cfg.output_folder = output_dir_backtrack
+    tmp_config = temp_directory / "config_rhine_pl.yaml"
     cfg.to_file(tmp_config)
     return str(tmp_config)
 
@@ -96,6 +129,14 @@ def preprocess(test_config_backward: str) -> None:
     runner = CliRunner()
     runner.invoke(
         cli, ["preprocess", "era5", test_config_backward], catch_exceptions=False
+    )
+
+
+@pytest.fixture(scope="session")
+def preprocess_pl(test_config_backward_pl: str) -> None:
+    runner = CliRunner()
+    runner.invoke(
+        cli, ["preprocess", "era5", test_config_backward_pl], catch_exceptions=False
     )
 
 
@@ -179,9 +220,7 @@ class TestPreprocess:
         output_path = preprocessed_dir / "2022-08-31_fluxes_storages.nc"
         assert output_path.exists()
         # verify outputs
-        expected_path = Path(
-            "tests/test_data/verify_output/preprocessed_data/2022-08-31_fluxes_storages.nc"
-        )
+        expected_path = Path(REGRESSION_DATA_MODEL_LAYERS)
         expected_output = xr.open_dataset(expected_path)
         output = xr.open_dataset(output_path)
 
@@ -196,6 +235,8 @@ class TestPreprocess:
             expected_output["precip"].values,
             output["precip"].values,
             err_msg=different_output_message(diff_figure, output_path, expected_path),
+            rtol=RTOL,
+            atol=ATOL,
         )
 
     def test_log_file(self, preprocessed_dir):
@@ -204,6 +245,38 @@ class TestPreprocess:
 
     def test_config(self, preprocessed_dir):
         config_path = preprocessed_dir / "config_rhine.yaml"
+        assert config_path.exists()
+
+
+class TestPreprocessPl:
+    @pytest.fixture(autouse=True)
+    def run_preprocess(self, preprocess_pl):
+        pass
+
+    def test_preprocess(self, preprocessed_dir_pl: Path):
+        output_path = preprocessed_dir_pl / "2020-01-01_fluxes_storages.nc"
+        assert output_path.exists()
+
+        # verify outputs
+        expected_path = Path(REGRESSION_DATA_PRESSURE_LAYERS)
+        expected_output = xr.open_dataset(expected_path)
+        output = xr.open_dataset(output_path)
+
+        for var in expected_output.data_vars:
+            numpy.testing.assert_allclose(
+                expected_output[var].values,
+                output[var].values,
+                err_msg=f"{var} not close enough.",
+                rtol=RTOL,
+                atol=ATOL,
+            )
+
+    def test_log_file(self, preprocessed_dir_pl):
+        log_files = [i for i in preprocessed_dir_pl.glob("wam2layers_*.log")]
+        assert len(log_files) >= 1
+
+    def test_config(self, preprocessed_dir_pl):
+        config_path = preprocessed_dir_pl / "config_rhine_pl.yaml"
         assert config_path.exists()
 
 
@@ -237,6 +310,8 @@ class TestBacktrack:
             expected_output["e_track"].values,
             output["e_track"].values,
             err_msg=different_output_message(diff_figure, output_path, expected_path),
+            rtol=RTOL,
+            atol=ATOL,
         )
 
     def test_time_coord(self, output_path: Path):
@@ -294,6 +369,8 @@ class TestForwardtrack:
                 err_msg=different_output_message(
                     diff_figure, output_path, expected_path
                 ),
+                rtol=RTOL,
+                atol=ATOL,
             )
 
     def test_time_coord(self, output_path: Path):