+++ CHANGELOG.md
@@ -4,13 +4,20 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+
+## Release v3.1.0 (2024-06-21)
+
 ### Changed
 
-- updated example config
+- The user guide has been restructured [#367](https://github.com/WAM2layers/WAM2layers/pull/367)
+- The example config has been updated [#371](https://github.com/WAM2layers/WAM2layers/pull/371)
+- Added config options to allow for specifying the layer boundary during preprocessing [#391](https://github.com/WAM2layers/WAM2layers/pull/391).
+- The preprocessing module has been restructured to allow for easier implementation of new datasets [#391](https://github.com/WAM2layers/WAM2layers/pull/391)
 
 ### Fixed
 
 - Fixed bug in treatment of periodic boundaries ([#379](https://github.com/WAM2layers/WAM2layers/pull/379)).
+- Pressure level (ERA5) data can be preprocessed again [#387](https://github.com/WAM2layers/WAM2layers/pull/387)
 
 ## Release v3.0.0 (2024-04-05)